chore: init router in main package
The goal of this commit is to start splitting handlers more
easily.
